+++ src/ECS/mod.rs
@@ -6,9 +6,11 @@ use super::*;
 
 mod comp;
 mod storage;
+mod system;
 
 use comp::*;
 use storage::*;
+use system::*;
 
 pub trait gmRes: Any{
     fn new() -> Self;
@@ -74,26 +76,6 @@ impl gmWorld{
                 COMP.downcast_mut::<&mut dyn gmStorageDrop>().unwrap().drop(IN_id);
             }
     }
-}
-
-pub trait gmSystem<'a>{
-    type sysData: gmSystemData<'a>;
-    fn new() -> Self;
-    fn SYS_ID() -> &'static str;
-    fn execute(&mut self, IN_data: Self::sysData);
-}
-
-pub trait gmSysRun<'a>{
-    fn executeNow(&mut self, IN_world: &'a mut gmWorld);
-}
-impl<'a, T> gmSysRun<'a> for T where T:gmSystem<'a>{
-    fn executeNow(&mut self, IN_world: &'a mut gmWorld) {
-        self.execute(T::sysData::fetch(IN_world));
-    }
-}
-
-pub trait gmSystemData<'a>{
-    fn fetch(IN_world: &'a mut gmWorld) -> Self;
 }
 
 pub struct gmDispatcher{
